Simulations are now (badly) calculated and returned! Whoo!
-Added Meters to GeoCoord's supported units
-Changed TelemDataPack to return int as totalVoltage (stored as int)
-Moved SimFrame to common. 
-added a bunch of stubby methods to add calculations.

+++ src/com/ubcsolar/common/TelemDataPacket.java
@@ -90,7 +90,7 @@ public Map<String, Object> getAllValues() {
 		return this.timeCreated;
 	}
 	
-	public float getTotalVoltage(){
+	public int getTotalVoltage(){
 		return totalVoltage;
 	}
